services/dle-ms/requirements.txt
flask
flask-cors
python-socketio[client]
